requirements.txt
asgiref==3.7.2
atomicwrites==1.4.0
attrs==21.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.4
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
idna==3.6
iniconfig==1.1.1
mssql-django==1.4.2
packaging==21.3
platformdirs==4.2.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.8.0
pyodbc==5.1.0
pyparsing==3.0.7
pytest==7.0.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
